Refactor authentication to remove redundant code for setting access_token_cookie.
The code in the authentication file was refactored to remove redundant code for setting the access_token_cookie. Previously, there was a conditional check to see if the auth object had a token property, and if so, set the access_token to the cookie. However, this conditional check was unnecessary because the token object was always present in the response. As a result, the conditional check was removed.