feat: Add debug flag to healthcheck response
This commit adds a debug flag to the healthcheck response. The flag indicates whether the application is running in debug mode or not. The flag is set to `True` if the `app.debug` property is `True`, otherwise it is set to `False`. This flag provides information about the application's debug status.

Also, update debug.ts to check for debug environment

This commit updates the `beforeSend` function in the `debug.ts` file to check if the application is running in debug mode. It does this by making a GET request to the `/api/health` endpoint and checking the value of the `debug` property in the response. If the `debug` property is `true`, the environment is set to 'debug', otherwise, it is set to 'production'. This change ensures that the correct environment is set when tracking page analytics using Matomo.