Delete encoders.py, README.md, create-migration.bat, and create-migration.sh.
This commit removes unnecessary files that are no longer needed in the codebase. The files include encoders.py, which contained classes for encoding and decoding datetime objects in JSON, README.md in the migrations folder, and create-migration.bat and create-migration.sh, which were used for creating migration scripts. These files are being deleted because they are no longer required and have been deemed unnecessary.