Update invite's "used_by" field with user id
Update invite's "used_by" field with user id

This commit updates the "used_by" field of the invite with the provided user id. It removes the user id from the "used_by" field if it exists, and appends the user id to the field if it doesn't already exist. If the "used_by" field becomes empty, it is set to None.