feat: Add GitHub Actions workflow for building and pushing Docker image
This commit adds a new GitHub Actions workflow file `docker-build-test.yml` which is triggered on a push to the `v3-beta` branch. The workflow sets up environment variables for the Docker registry and image name, and runs a job to build and push the Docker image. It includes steps to checkout the repository, cache Docker layers, set up QEMU and enable BuildKit, log into the registry, extract metadata for Docker, set up Docker Buildx, and finally build and push the Docker image. Additionally, there is an update to the Dockerfile to replace `npm ci` with `npm install`.