Update tags in docker-build.yml
This commit updates the tags for Docker images in the docker-build.yml file to include the registry information. It replaces the previous tags with the new format: ${{ env.REGISTRY }}:${{ env.IMAGE_NAME }}:${{ env.IMAGE_TAG }}-${{ matrix.platform }}. This change ensures that the correct registry is used when pushing the images.