feat: Add Discord component for displaying Discord server members
This commit adds a new component called Discord.vue to the help module. The Discord component is responsible for displaying the members of a Discord server. It includes a header that shows the number of online members, a members list with avatars and usernames, and a footer with a link to join the Discord server. The component also makes use of the Discord Widget API to retrieve the member data.