feat: Update Docker Build and Push descriptions
Updated the names of the workflows to accurately reflect their purpose - the "name" field was changed for the "Docker Build and Push" and "Release Docker Image CI" workflows in the `beta-ci.yml` and `ci.yml` files respectively. The updated names now clearly indicate whether the workflow is for the beta or release builds.