Refactor CI workflow with check_changes job and dependencies
Previously, the CI workflow did not check if changes only applied to README.md, .github, .vscode, resulting in unnecessary builds. With this commit, a new job named "check_changes" has been added to ensure that the workflow does not run if changes are limited to these files. The "build_ghcr" and "build_dh" jobs have also been updated to depend on the "check_changes" job. This improvement helps optimize the CI process and saves resources.