feat: Add support for syncing users from Jellyfin server
This commit introduces a code change that adds support for syncing users from a Jellyfin media server. Previously, users were only synced from Plex servers. With this change, the `global_sync_users_to_media_server` and `global_get_user_profile_picture` functions now check for the server type and invoke the respective functions for syncing and retrieving user information based on the server type.

The code in the `global_invite_user_to_media_server` function was also updated to include the creation timestamp when adding a new user to the database.

---

chore: Update Docker images with additional packages

This commit updates the Dockerfiles for the backend, frontend, and CI containers. The `wizarr-backend`, `wizarr-frontend`, and `wizarr-ci` containers now include the `figlet` package in addition to the existing packages. The `figlet` package is necessary for generating ASCII art in certain components of the application.

The `wizarr` container was also updated to include the `figlet` package.