Add RequestsList component
This commit adds a new file `RequestsList.vue` which defines the `RequestsList` component. This component displays a list of requests with their names and icons. Each request item also has a button to open the request URL in a new tab. The component uses the `ListItem` component to render each request item. The requests and their icons are loaded dynamically using the `requestURLS` prop.