Refactor support page, add toast notification for session end.
The code changes in the `Support.vue` file refactor the support page by adding a new import and a new variable `session_toast` of type `ToastID` from the `vue-toastification` library. The `sessionFinished` method is modified to dismiss the `session_toast` if it exists, and then create a new toast for the end of the support session. The `startSupportSession` method is modified to conditionally set the guest details using the `support` object based on the presence of certain properties.

The code changes in the `LoadingWidget.vue` file modify the formatting of the code to remove unnecessary line breaks and indentation.

The code changes in the `LatestInfo.vue` file update the `readMore` method to open a modal window with the `newLineExplicit` computed property as the content. The `latestInfo` data property is updated to be either `null` or a string value.