chore: Update version to 3.4.2 and add automated version bumping
The commit updates the version in the package.json file to 3.4.2. Additionally, it adds an automated workflow to bump the version to "pre-beta" using the `phips28/gh-action-bump-version` action. The `latest` file is also updated to match the package.json version using `jq`, and these changes are committed using the `stefanzweifel/git-auto-commit-action`.