ci & release workflow updates: add new trigger & default input
This commit modifies the CI and release workflows to include a new trigger for when a release is published. Additionally, a new default input is added to the workflow_dispatch trigger, allowing the version to be specified when building and pushing. The dependency installation step in the release workflow is updated to ignore scripts. The version in the package.json file is also updated.