feat: Build frontend sourcemaps for beta release
Prior to the beta release, this commit adds a step to build the frontend sourcemaps using the command `npx nx build wizarr-frontend`. This ensures that the sourcemaps for the frontend are generated and ready for release.

In addition, the commit configures the semantic release process to include the frontend sourcemaps in the release. The `@wizarrrr/semantic-release-sentry-releases` plugin is modified to specify the location of the sourcemaps as `dist/apps/wizarr-frontend`.