Fix dark mode toggle
The dark mode toggle would sometimes not work on load (going back and forth on the pages makes it happen pretty often). This fixes the issue by leveraging the hx-on:load and hx-on:click to call the javascript functions. This eliminates the issue with the click listener not working right with the content being swapped.